Implement order entity

+++ src/main/java/com/app/onlinebookstore/service/impl/CartItemServiceImpl.java
@@ -50,6 +50,7 @@ public class CartItemServiceImpl implements CartItemService {
     }
 
     @Override
+    @Transactional
     public CartItemDto update(Long id, UpdateCartItemDto updateCartItemDto) {
         CartItem cartItem = cartItemRepository.findById(id).orElseThrow(
                 () -> new EntityNotFoundException("Can't find item by id " + id));
@@ -66,7 +67,16 @@ public class CartItemServiceImpl implements CartItemService {
 
     @Override
     public void delete(Long id) {
-        cartItemRepository.deleteById(id);
+        Long userId = userService.getAuthenticatedUser().getId();
+        Optional<ShoppingCart> byUserId = shoppingCartRepository.findByUserId(userId);
+        List<CartItem> cartItemsByShoppingCartId = cartItemRepository
+                .findCartItemsByShoppingCartId(byUserId.get().getId());
+        boolean exist = cartItemsByShoppingCartId.stream()
+                .map(CartItem::getId)
+                .anyMatch(i -> i.equals(id));
+        if (exist) {
+            cartItemRepository.deleteById(id);
+        }
     }
 
     private void setShoppingCartAndCartItems(Long shoppingCartId, CartItem cartItem) {
